Changes in Payments, Orders, Create order, Referral request
Payments: The payments route was sending just the list of payments and was not providing the user's information on tier and first rent payment. Orders : Create order route not longer pushes the payment created to the user in a payment list. Not handles when a user (reference) registered a wrong referral. Create order: Now referrals payment function sends the reference referall to query the user to pay instead of user id to fix the problem of not find the user by id and throwing an error instead of null value. Referral request: The query was changed to query first the user and then reference instead of populating the reference of a user from a list. This is done to avoid redundancy